requirements.txt
streamlit
tensorflow
pillow
numpy
opencv-python
matplotlib
scikit-learn
pandas